requirements.txt
streamlit_webrtc==0.47.1
twilio~=8.9.1
streamlit
pyyaml
cloudinary
numpy
python-dotenv